Don't set RXEN on RX FIFO slave objects
This is not necessary for proper RX FIFO operation and it results in
possible reception of frames in the middle of the RX FIFO because
RX FIFO slave objects are normal message objects that take part in
the acceptance filtering process if RXEN is set.

This is especially noticable when using a read filter, because then
frames that would be rejected by the read filter are recevied into
the RX FIFO slave objects messing with the operation of the RX FIFO
resulting in receiving unwanted frames out-of-order.